requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
